Distributed XGBoost scoring which handles rebalanced data
